pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "wedley", email = ""},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm]
